Increase coverage of shortest path solver integration test
By using doubles, the solvers would often be able to short-circuit
which, while nice in practice, makes it easier for the test to pass than
we'd like. We fix this here by replacing the ensuring that inputs are
square, by replacing the random doubles with random integers, and by
replacing the assignment equality test with an objective equality test.